# apps/api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Gonepost API
  version: 0.1.0
  description: >
    API contract for the basecode project. This spec is the source of truth:
    Go server types and the TypeScript client are both generated from it.
    Domain endpoints are versioned under /api/v1; infrastructure endpoints
    used by orchestrators/load balancers (health, readiness) are kept
    unversioned since they are polled directly and never change shape.

security:
  - sessionCookie: []

paths:
  /health:
    get:
      operationId: getHealth
      summary: Liveness probe
      description: Returns 200 if the process is up. Does not check dependencies.
      tags: [system]
      security: []
      responses:
        "200":
          description: Service is alive
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /ready:
    get:
      operationId: getReady
      summary: Readiness probe
      description: >
        Returns 200 if the process is ready to accept traffic. Will later
        check dependencies (database, etc.) before returning ready.
      tags: [system]
      security: []
      responses:
        "200":
          description: Service is ready
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/v1/auth/login:
    post:
      operationId: login
      summary: Log in with email and password
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in — session cookie set via Set-Cookie
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Account is not active
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
        "429":
          description: Too many login attempts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/v1/auth/logout:
    post:
      operationId: logout
      summary: Log out and revoke the current session
      tags: [auth]
      responses:
        "204":
          description: Logged out

  /api/v1/auth/me:
    get:
      operationId: getCurrentUser
      summary: Get the currently authenticated user
      tags: [auth]
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/v1/auth/bootstrap-admin:
    post:
      operationId: bootstrapSuperAdmin
      summary: Grant SUPER_ADMIN to the calling user (only works once)
      description: >
        Only succeeds if no user currently holds SUPER_ADMIN. Requires being
        logged in — there is no permission check, since a brand-new
        deployment has no super admin yet to have granted one.
      tags: [auth]
      responses:
        "204":
          description: Bootstrapped — caller is now SUPER_ADMIN
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: A super admin already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/v1/auth/password/change:
    post:
      operationId: changePassword
      summary: Change the current user's password
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "204":
          description: Password changed
        "401":
          description: Not authenticated or old password incorrect
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"

  /api/v1/users:
    get:
      operationId: listUsers
      summary: List users (paginated, filterable by status)
      tags: [users]
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/UserStatus"
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Paginated user list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
    post:
      operationId: createUser
      summary: Create a user
      tags: [users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "409":
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"

  /api/v1/users/{id}:
    get:
      operationId: getUser
      summary: Get a user by ID
      tags: [users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateUser
      summary: Update a user's profile fields and status
      tags: [users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"

  /api/v1/users/{id}/roles:
    get:
      operationId: listUserRoles
      summary: List roles held by a user
      tags: [roles]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Roles held by the user
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Role"
    post:
      operationId: assignRoleToUser
      summary: Assign a role to a user
      tags: [roles]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignRoleRequest"
      responses:
        "204":
          description: Assigned

  /api/v1/users/{id}/roles/{roleId}:
    delete:
      operationId: removeRoleFromUser
      summary: Remove a role from a user
      description: >
        Refuses (409) if this would remove the last active SUPER_ADMIN.
      tags: [roles]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: roleId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Removed
        "409":
          description: Would remove the last active super admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/v1/roles:
    get:
      operationId: listRoles
      summary: List roles
      tags: [roles]
      responses:
        "200":
          description: Roles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Role"
    post:
      operationId: createRole
      summary: Create a role
      tags: [roles]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRoleRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"

  /api/v1/roles/{id}/permissions:
    get:
      operationId: listRolePermissions
      summary: List permission codes granted to a role
      tags: [roles]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Permission codes
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
    post:
      operationId: grantPermissionToRole
      summary: Grant a permission to a role
      tags: [roles]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GrantPermissionRequest"
      responses:
        "204":
          description: Granted
        "404":
          description: Permission code not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"


  /api/v1/roles/{id}/permissions/{code}:
    delete:
      operationId: revokePermissionFromRole
      summary: Revoke a permission from a role
      tags: [roles]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Revoked
        "404":
          description: Permission code not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/v1/audit-logs:
    get:
      operationId: listAuditLogs
      summary: List audit events (paginated, filterable)
      tags: [auditlog]
      parameters:
        - name: action
          in: query
          schema:
            type: string
        - name: actor_user_id
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Paginated audit events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditLogListResponse"

  /api/v1/auth/permissions:
    get:
      operationId: getMyPermissions
      summary: Permission codes the current user holds (for showing menus and guarding pages)
      tags: [auth]
      responses:
        "200":
          description: Permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PermissionList"
        "401":
          description: Not logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/v1/items:
    get:
      operationId: listItems
      summary: List the caller's items (paginated, searchable, filterable by status). Needs items:read.
      tags: [items]
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/ItemStatus"
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Paginated item list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemListResponse"
    post:
      operationId: createItem
      summary: Create an item. Needs items:write.
      tags: [items]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"

  /api/v1/items/{id}:
    get:
      operationId: getItem
      summary: Get one of the caller's items. Needs items:read.
      tags: [items]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateItem
      summary: Replace an item's fields. Needs items:write.
      tags: [items]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItemInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
    delete:
      operationId: deleteItem
      summary: Delete an item. Needs items:write.
      tags: [items]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session_token

  schemas:
    StatusResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
          example: ok

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: internal_error
            message:
              type: string
              example: something went wrong

    ValidationErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, fields]
          properties:
            code:
              type: string
              example: validation_error
            fields:
              type: array
              items:
                type: object
                required: [field, message]
                properties:
                  field:
                    type: string
                  message:
                    type: string

    UserStatus:
      type: string
      enum: [active, inactive, suspended]

    User:
      type: object
      required: [id, email, status, created_at, updated_at]
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        username:
          type: string
        full_name:
          type: string
        status:
          $ref: "#/components/schemas/UserStatus"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    UserListResponse:
      type: object
      required: [data, total, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        limit:
          type: integer

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    ChangePasswordRequest:
      type: object
      required: [old_password, new_password]
      properties:
        old_password:
          type: string
        new_password:
          type: string
          minLength: 8

    CreateUserRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        username:
          type: string
        full_name:
          type: string

    UpdateUserRequest:
      type: object
      required: [status]
      properties:
        username:
          type: string
        full_name:
          type: string
        status:
          $ref: "#/components/schemas/UserStatus"

    Role:
      type: object
      required: [id, name, is_system]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        is_system:
          type: boolean

    CreateRoleRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string

    AssignRoleRequest:
      type: object
      required: [role_id]
      properties:
        role_id:
          type: string

    GrantPermissionRequest:
      type: object
      required: [code]
      properties:
        code:
          type: string
          example: users:read

    AuditLogEvent:
      type: object
      required: [id, action, resource, result, created_at]
      properties:
        id:
          type: string
        actor_user_id:
          type: string
        action:
          type: string
        resource:
          type: string
        resource_id:
          type: string
        result:
          type: string
          enum: [success, failure]
        request_id:
          type: string
        metadata:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time

    AuditLogListResponse:
      type: object
      required: [data, total, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AuditLogEvent"
        total:
          type: integer
        limit:
          type: integer

    PermissionList:
      type: object
      required: [permissions]
      properties:
        permissions:
          type: array
          items:
            type: string
          example: ["items:read", "items:write"]

    ItemStatus:
      type: string
      enum: [active, archived]

    Item:
      type: object
      required: [id, name, status, created_at, updated_at]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        status:
          $ref: "#/components/schemas/ItemStatus"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    ItemInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        status:
          $ref: "#/components/schemas/ItemStatus"

    ItemListResponse:
      type: object
      required: [data, total, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Item"
        total:
          type: integer
        limit:
          type: integer